Split runtime tests into runtime/tests/{support,behavior_resolution,router,control_watcher,startup_recovery}.rs
Converts the 1386-line tests.rs into a tests/ directory with 5 sibling files:
  mod.rs (8 lines) — sub-module declarations and re-exports only
  support.rs (555 lines) — shared helpers and fixtures; soft-cap fallback applied
    (555 lines exceeds the 400-line soft target and the 500-line warning threshold
    but support helpers are tightly coupled and cannot be split further)
  behavior_resolution.rs (65 lines) — 3 resolve_behavior_* tests
  router.rs (146 lines) — 2 router_* tests
  control_watcher.rs (252 lines) — 3 control_watcher_* tests + local helper
  startup_recovery.rs (375 lines) — 5 run_agent_* tests + local helper

All 13 tests pass. Test names are unchanged except for the added submodule
path segment (e.g. agent::runtime::tests::router::router_dispatches_*).
