+++ crates/defra-agent/src/agent/runtime/tests/mod.rs
@@ -1,0 +1,8 @@
+mod support;
+mod behavior_resolution;
+mod control_watcher;
+mod router;
+mod startup_recovery;
+
+pub(super) use support::*;
+pub(super) use super::*;
